Split iOS workspace linking responsibilities
